requirements.txt
# --- Core web/app ---
Flask>=3.0
python-dotenv>=1.0
# Optional WSGI server for Linux/macOS (ignored on Windows):
gunicorn>=21.2 ; platform_system != "Windows"

# --- Imaging & utilities ---
Pillow>=10.0,<11
requests>=2.31
numpy>=1.23,<2.0
pandas>=2.2
matplotlib>=3.8
opencv-python>=4.8

# --- ML / CV ---
torch>=2.3
torchvision>=0.18
timm>=0.9.12
pytorch-grad-cam>=1.5.0
scikit-learn>=1.3,<1.6
ultralytics>=8.2.0

# If you need a CUDA build of PyTorch, install it separately, e.g.:
# pip install --index-url https://download.pytorch.org/whl/cu121 torch torchvision
